requirements.txt
oandapyV20==0.7.2
streamlit>=1.35.0
plotly>=5.22.0
streamlit-autorefresh>=1.0.1
anthropic>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
ta>=0.11.0
yfinance>=0.2.40
groq>=0.11.0
google-genai>=1.0.0
google-cloud-firestore>=2.16.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
